+++ gradle.properties
@@ -19,6 +19,7 @@ log4jVersion=2.17.1
 forgeAutoRenamingToolVersion=0.1.24
 mixinVersion=0.8.5
 modlauncherVersion=8.1.3
+java9hacksVersion=1.3.3
 guavaVersion=21.0
 junitVersion=5.9.1
 mockitoVersion=4.8.0
