+++ src/launch/java/org/spongepowered/common/launch/plugin/VanillaDummyPluginContainer.java
@@ -22,7 +22,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package org.spongepowered.vanilla.launch.plugin;
+package org.spongepowered.common.launch.plugin;
 
 import com.google.common.base.MoreObjects;
 import org.apache.logging.log4j.Logger;
