+++ src/test/resources/META-INF/services/cpw.mods.modlauncher.api.ILaunchHandlerService
@@ -1,0 +1,1 @@
+org.spongepowered.common.launch.TestLaunchHandler
